One pk must be called pk
Apparently DRF requires one of the ids in the URL to be called pk. Or
you can set .lookup_field or something, but I couldn't get that to work.
This was the error:

```
AssertionError at /teams/1/trainers/
Expected view TrainerView to be called with a URL keyword argument named
"pk". Fix your URL conf, or set the `.lookup_field` attribute on the
view correctly.
```